+++ src/app/appEvents.ts
@@ -5,16 +5,14 @@ import { clamp } from "../math/clamp";
 import { _k } from "../shared";
 import { toFixed } from "../utils/numbers";
 
-// Events used at the start of a game
-
 export function initAppEvents() {
-    _k.app.onHide(() => {
+    _k.appScope.onTabHide(() => {
         if (!_k.globalOpt.backgroundAudio) {
             _k.audio.ctx.suspend();
         }
     });
 
-    _k.app.onShow(() => {
+    _k.appScope.onTabShow(() => {
         if (!_k.globalOpt.backgroundAudio && !_k.debug.paused) {
             _k.audio.ctx.resume();
         }
@@ -49,29 +47,29 @@ export function initAppEvents() {
     });
 
     if (_k.globalOpt.debug !== false) {
-        _k.app.onKeyPress(
+        _k.appScope.onKeyPress(
             _k.globalOpt.debugKey ?? "f1",
             () => _k.debug.inspect = !_k.debug.inspect,
         );
-        _k.app.onKeyPress("f2", () => _k.debug.clearLog());
-        _k.app.onKeyPress("f8", () => _k.debug.paused = !_k.debug.paused);
-        _k.app.onKeyPress("f7", () => {
+        _k.appScope.onKeyPress("f2", () => _k.debug.clearLog());
+        _k.appScope.onKeyPress("f8", () => _k.debug.paused = !_k.debug.paused);
+        _k.appScope.onKeyPress("f7", () => {
             _k.debug.timeScale = toFixed(
                 clamp(_k.debug.timeScale - 0.2, 0, 2),
                 1,
             );
         });
-        _k.app.onKeyPress("f9", () => {
+        _k.appScope.onKeyPress("f9", () => {
             _k.debug.timeScale = toFixed(
                 clamp(_k.debug.timeScale + 0.2, 0, 2),
                 1,
             );
         });
-        _k.app.onKeyPress("f10", () => _k.debug.stepFrame());
+        _k.appScope.onKeyPress("f10", () => _k.debug.stepFrame());
     }
 
     // burp mode initialization
     if (_k.globalOpt.burp) {
-        _k.app.onKeyPress("b", () => burp());
+        _k.appScope.onKeyPress("b", () => burp());
     }
 }