+++ src/core/engine.ts
@@ -23,6 +23,12 @@ import { rotateFactory } from "../ecs/components/transform/rotate";
 import { scaleFactory } from "../ecs/components/transform/scale";
 import { zFactory } from "../ecs/components/transform/z";
 import { registerPrefabFactory } from "../ecs/entity/prefab";
+import { createGameEventHandlers } from "../events/gameEventHandlers";
+import {
+    attachAppHandlersToGameObjRaw,
+    createAppScope,
+    createSceneScope,
+} from "../events/scopes";
 import { createGame } from "../game/game";
 import { createCanvas } from "../gfx/canvas";
 import { initGfx } from "../gfx/gfx";
@@ -53,6 +59,7 @@ export const createEngine = (gopt: KAPLAYOpt) => {
         {
             scale: 1,
             spriteAtlasPadding: 2,
+            defaultLifetimeScope: "scene" as "scene" | "app",
         } satisfies KAPLAYOpt,
         gopt,
     );
@@ -60,6 +67,10 @@ export const createEngine = (gopt: KAPLAYOpt) => {
     const canvas = createCanvas(opt);
     const { fontCacheC2d, fontCacheCanvas } = createFontCache();
     const app = initApp({ canvas, ...gopt });
+    const gameHandlers = createGameEventHandlers(app);
+    const sceneScope = createSceneScope(app, gameHandlers);
+    const appScope = createAppScope(gameHandlers);
+    attachAppHandlersToGameObjRaw(gameHandlers);
 
     // TODO: Probably we should move this to initGfx
     const canvasContext = app.canvas
@@ -84,6 +95,7 @@ export const createEngine = (gopt: KAPLAYOpt) => {
 
     // Frame rendering
     const frameRenderer = createFrameRenderer(
+        app,
         appGfx,
         game,
         opt.pixelDensity ?? 1,
@@ -139,6 +151,8 @@ export const createEngine = (gopt: KAPLAYOpt) => {
         game,
         debug,
         gc: [] as (() => void)[],
+        sceneScope,
+        appScope,
         // Patch, k it's only avaible after running kaplay()
         k: null as unknown as KAPLAYCtx,
         startLoop() {