+++ src/ecs/entity/GameObjRaw.ts
@@ -9,6 +9,7 @@ import {
     KEventController,
     type KEventHandler,
 } from "../../events/events";
+import type { GameEventHandlers } from "../../events/gameEventHandlers";
 import {
     onAdd,
     onDestroy,
@@ -481,6 +482,10 @@ export interface GameObjRaw {
     onButtonDown: KAPLAYCtx["onButtonDown"];
     onButtonPress: KAPLAYCtx["onButtonPress"];
     onButtonRelease: KAPLAYCtx["onButtonRelease"];
+    onTabShow: GameEventHandlers["onTabShow"];
+    onTabHide: GameEventHandlers["onTabHide"];
+    onShow: GameEventHandlers["onShow"];
+    onHide: GameEventHandlers["onHide"];
 }
 
 export type InternalGameObjRaw = GameObjRaw & {
@@ -560,7 +565,11 @@ const COMP_EVENTS = new Set([
 
 type GarbageCollectorArray = (() => any)[];
 
-export const GameObjRawPrototype: Omit<InternalGameObjRaw, AppEvents> = {
+export const GameObjRawPrototype: Omit<
+    InternalGameObjRaw,
+    // TODO: Maybe too hacky, find better way
+    Exclude<AppEvents, "onFixedUpdate" | "onUpdate" | "onDraw">
+> = {
     // This chain of `as any`, is because we never should use this object
     // directly, it's only a prototype. These properties WILL be defined
     // (by our factory function `make`) when we create a new game object.
@@ -1563,64 +1572,3 @@ export const GameObjRawPrototype: Omit<InternalGameObjRaw, AppEvents> = {
     },
     // #endregion
 };
-
-// #region App Events in Proto
-export function attachAppToGameObjRaw() {
-    // We add App Events for "attaching" it to game object
-    const appEvs = [
-        "onKeyPress",
-        "onKeyPressRepeat",
-        "onKeyDown",
-        "onKeyRelease",
-        "onMousePress",
-        "onMouseDown",
-        "onMouseRelease",
-        "onMouseMove",
-        "onCharInput",
-        "onMouseMove",
-        "onTouchStart",
-        "onTouchMove",
-        "onTouchEnd",
-        "onScroll",
-        "onGamepadButtonPress",
-        "onGamepadButtonDown",
-        "onGamepadButtonRelease",
-        "onGamepadStick",
-        "onButtonPress",
-        "onButtonDown",
-        "onButtonRelease",
-    ] satisfies [...AppEvents[]];
-
-    for (const e of appEvs) {
-        const obj = GameObjRawPrototype as Record<string, any>;
-
-        obj[e] = function(this: InternalGameObjRaw, ...args: [any]) {
-            // @ts-ignore
-            const ev: KEventController = _k.app[e]?.(...args);
-            ev.paused = this.paused;
-
-            this._inputEvents.push(ev);
-
-            this.onDestroy(() => ev.cancel());
-
-            // This only happens if obj.has("stay");
-            this.on("sceneEnter", () => {
-                // All app events are already canceled by changing the scene
-                // so we don't need to event.cancel();
-                this._inputEvents.splice(this._inputEvents.indexOf(ev), 1);
-
-                // create a new event with the same arguments
-                // @ts-ignore
-                const newEv = _k.app[e]?.(...args);
-
-                // Replace the old event handler with the new one
-                // old KEventController.cancel() => new KEventController.cancel()
-                KEventController.replace(ev, newEv);
-                this._inputEvents.push(ev);
-            });
-
-            return ev;
-        };
-    }
-}
-// #endregion