+++ src/types.ts
@@ -135,12 +135,21 @@ export type Key =
     | (string & {});
 
 /**
+ * You can use 3 or more keys (like `control+shift+s`),
+ * it just isn't included in the type here because typescript
+ * crashes when it tries to expand all 74^3 = 405224 combinations
+ * for 3 keys.
+ */
+export type ChordedKey = Key | `${Key}+${Key}`;
+
+/**
  * A mouse button.
  *
  * @group Input
  * @subgroup Mouse
  */
 export type MouseButton = "left" | "right" | "middle" | "back" | "forward";
+export type ChordedMouseButton = MouseButton | `${MouseButton}+${MouseButton}`;
 
 /**
  * A gamepad button.
@@ -168,6 +177,10 @@ export type KGamepadButton =
     | "home"
     | "capture"
     | "touchpad";
+export type ChordedKGamepadButton =
+    | KGamepadButton
+    | `${KGamepadButton}+${KGamepadButton}`
+    | `${KGamepadButton}+${KGamepadButton}+${KGamepadButton}`;
 
 /**
  * A gamepad stick.
@@ -205,6 +218,9 @@ export type KGamepad = {
     isReleased(b: KGamepadButton): boolean;
     /** Get the value of a stick. */
     getStick(stick: KGamepadStick): Vec2;
+    /** Get the 0-1 analog value of the button
+     * (useful for `ltrigger` and `rtrigger` buttons) */
+    getAnalog(b: KGamepadButton): number;
 };
 
 /**
@@ -382,6 +398,12 @@ export interface KAPLAYOpt {
      * @default 3000
      */
     loadTimeout?: number;
+    /**
+     * The default lifetime scope used for event handlers.
+     *
+     * @default "scene"
+     */
+    defaultLifetimeScope?: "scene" | "app";
     /**
      * TypeScript Advanced Features (TAF) are a series of options for TypeScript
      * only features.