+++ tests/playtests/keyandkeycode.js
@@ -1,0 +1,16 @@
+kaplay({
+    buttons: {
+        foo: {
+            keyboard: "q",
+            keyboardCode: "KeyQ",
+        },
+    },
+});
+
+onButtonPress("foo", () => {
+    debug.log("ohhi");
+});
+
+onButtonRelease("foo", () => {
+    debug.log("ohbye");
+});